Major improvements and additions
Add several new powerups
Improve AnimationSystem
Improve RenderSystem
Fix bugs in RenderSystem
Many other changes and improvements

+++ core/src/no/ntnu/tdt4240/asteroids/entity/component/BoundsComponent.java
@@ -31,7 +31,7 @@ public abstract class BoundsComponent implements Component, Pool.Poolable {
 
     public abstract void setCenter(Vector2 vector);
 
-    public abstract void setSize(int width, int height);
+    public abstract void setSize(float width, float height);
 
     public abstract Vector2 getCenter(Vector2 result);
 }
